Fix scriptless behaviors
At present, if an entity has a defined behavior without a defined script, we will return from the behavior attachment method. This means that behaviors defined afterwards are simply not attached, regardless of whether or not they have scripts. 

This fix means that we can both use behavior definitions as metadata (As seen in the room-commands implementation in base Ranvier), and prevent hard-to-track bugs caused by defined behaviors not being attached without warning).